FINISHED - I think.
Added comments for each of the functions, explaining what they do and what arguments they need etc. Also improved the redundancy of the program, made it so it can tell you when you cannot save to a file if you dont have any data to store and when you cannot load from a file if it does not exist.